Support Emscripten EH/SjLj in Wasm64
This reflects changes in the function signature in
https://reviews.llvm.org/D101985 and also uses `from64` function defined
in #12869. This has not been tested against wasm64 yet and I'm uploading
this mainly for discussions at this point.

Also in the team chat of the tools team we discussed other alternatives
to `from64` function, which can be done separately and not reflected
here.

+++ system/lib/compiler-rt/emscripten_exception_builtins.c
@@ -11,10 +11,10 @@
 
 #include <threads.h>
 
-thread_local int __THREW__ = 0;
+thread_local uintptr_t __THREW__ = 0;
 thread_local int __threwValue = 0;
 
-void setThrew(int threw, int value) {
+void setThrew(uintptr_t threw, int value) {
   if (__THREW__ == 0) {
     __THREW__ = threw;
     __threwValue = value;
